requirements.txt
Flask==2.0.2
openai==0.26.5
requests_toolbelt==0.10.1
Pillow==9.5.0